+++ A2Z/Form1.Designer.cs
@@ -82,6 +82,7 @@ namespace A2Z
             this.btnGenerate2D = new System.Windows.Forms.Button();
             this.btnClashDetection = new System.Windows.Forms.Button();
             this.btnCollectBOM = new System.Windows.Forms.Button();
+            this.btnExcelTemplatePoC = new System.Windows.Forms.Button();
             this.btnMainDimension = new System.Windows.Forms.Button();
             this.panelGlobalViewButtons = new System.Windows.Forms.Panel();
             this.panelGlobalActions = new System.Windows.Forms.Panel();
@@ -623,12 +624,13 @@ namespace A2Z
             this.groupBox1.Controls.Add(this.btnGenerate2D);
             this.groupBox1.Controls.Add(this.btnClashDetection);
             this.groupBox1.Controls.Add(this.btnCollectBOM);
+            this.groupBox1.Controls.Add(this.btnExcelTemplatePoC);
             this.groupBox1.Dock = System.Windows.Forms.DockStyle.Top;
             this.groupBox1.Location = new System.Drawing.Point(3, 3);
             this.groupBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             this.groupBox1.Name = "groupBox1";
             this.groupBox1.Padding = new System.Windows.Forms.Padding(3, 4, 3, 4);
-            this.groupBox1.Size = new System.Drawing.Size(443, 55);
+            this.groupBox1.Size = new System.Drawing.Size(530, 55);
             this.groupBox1.TabIndex = 0;
             this.groupBox1.TabStop = false;
             this.groupBox1.Text = "작업";
@@ -704,6 +706,18 @@ namespace A2Z
             this.btnExportPDF.Text = "PDF 내보내기";
             this.btnExportPDF.UseVisualStyleBackColor = true;
             this.btnExportPDF.Click += new System.EventHandler(this.btnExportPDF_Click);
+            //
+            // btnExcelTemplatePoC (REQ-002 / T-012 PoC Step 1)
+            //
+            this.btnExcelTemplatePoC.Font = new System.Drawing.Font("맑은 고딕", 8F);
+            this.btnExcelTemplatePoC.Location = new System.Drawing.Point(446, 20);
+            this.btnExcelTemplatePoC.Margin = new System.Windows.Forms.Padding(1);
+            this.btnExcelTemplatePoC.Name = "btnExcelTemplatePoC";
+            this.btnExcelTemplatePoC.Size = new System.Drawing.Size(78, 25);
+            this.btnExcelTemplatePoC.TabIndex = 11;
+            this.btnExcelTemplatePoC.Text = "엑셀 PoC";
+            this.btnExcelTemplatePoC.UseVisualStyleBackColor = true;
+            this.btnExcelTemplatePoC.Click += new System.EventHandler(this.btnExcelTemplatePoC_Click);
             //
             // btnMainDimension
             // 
@@ -1321,6 +1335,7 @@ namespace A2Z
         private System.Windows.Forms.Button btnExportPDF;
         private System.Windows.Forms.Button btnCollectOsnap;
         private System.Windows.Forms.Button btnExtractDimension;
+        private System.Windows.Forms.Button btnExcelTemplatePoC;
         private System.Windows.Forms.GroupBox groupBox5;
         private System.Windows.Forms.ListView lvDimension;
         private System.Windows.Forms.ColumnHeader colDimNo;